chore[cli]: Renamed to kebab case for standard
